requirements.txt
# Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
mysqlclient>=2.2.0

# File Storage
Pillow>=10.0.0
django-storages>=1.14.2
boto3>=1.28.62

# Authentication
djangorestframework-simplejwt>=5.3.0

# API Documentation
drf-yasg>=1.21.5
django-filter>=23.3

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
python-decouple>=3.8

# Email
sendgrid>=6.11.0
python-http-client>=3.3.7

# SMS (Twilio)
twilio>=8.9.0

# Payments
stripe>=7.4.0
razorpay>=1.3.0

# Caching
redis>=3.5.0,<4.0.0
redis-py-cluster>=2.1.3

# Monitoring
sentry-sdk>=1.34.0

# Security
django-axes>=6.0.3
django-ratelimit>=3.0.1
